US Debt Tracker Project
This dataset includes 20 years of data on Debt Held by the Public, Intragovernmental Holdings and Total Public Debt Outstanding. There were three questions to answer for this project.

Question 1: What was the Yearly Debt Percentage Increase for each year compared to the previous year?
Question 2: Which months historically have seen the highest/lowest increases in Total debt?
Question 3: What is the projected growth of the publicly held debt in the next few years?

Started the project off by making sure to clean the dataset. First, the rows with blank data were removed. Then Null values were replaced with 0s. To make the debt amounts easier to read, it was changed from scientific notation to number. 

Question 1: What was the Yearly Debt Percentage Increase for each year compared to the previous year?
Found the debt amount at the end of each year (last day of December of each year) in order to get the yearly debt percentage for each year. A simple filter showed these amounts. Then subtracted the debt amount of the current year and the previous year, this amount was divided by the previous year debt amount and then multiplied by 100 to give the percentage increase per year. A line chart was created showcase percentage increase per year for each debt type between 2016 - 2022. 

Question 2: Which months historically have seen the highest/lowest increases in Total debt?
Created a pivot table to find the average of total public debt outstanding for each month. A bar chart was created to show the average debt amount for each month. 

Question 3: What is the projected growth of the publicly held debt in the next few years?
A forecast function based on the maximum debt per year was used to predict the projected growth. A stacked area chart was created to show the projected growth of the publicly held debt in the next few years.